java, algos, ml, despair

+++ C++/CP/template.cpp
@@ -50,9 +50,9 @@ using namespace std;
 #define all(x)          (x).begin(), (x).end()
 
 // looping
-#define FOR(i, a, b)    for (int i = (a); i < (b); ++i)
+#define FOR(i, a, b)    for (ll i = (a); i < (b); ++i)
 #define F0R(i, a)       FOR(i, 0, a)
-#define ROF(i, a, b)    for (int i = (b)-1; i >= (a); --i)
+#define ROF(i, a, b)    for (ll i = (b)-1; i >= (a); --i)
 #define R0F(i, a)       ROF(i, 0, a)
 
 // general
